+++ Encoding/SvtAv1.vb
@@ -122,7 +122,7 @@ Public Class SVTAV1
                     Add("Basic",
                         New StringParam With {.Text = "Custom", .Quotes = QuotesMode.Never, .AlwaysOn = True},
                         Mode,
-                        New OptionParam With {.Switch = "-enc-mode", .Text = "Preset", .Init = 8, .IntegerValue = True, .Options = {"Very Slow", "Slower", "Slow", "Medium", "Fast", "Faster", "Very Fast", "Super Fast", "Ultra Fast"}},
+                        New OptionParam With {.Switch = "-enc-mode", .Text = "Preset", .Init = 8, .IntegerValue = True, .Options = {"0: Very Slow", "1: Slower", "2: Slow", "3: Medium", "4: Fast", "5: Faster", "6: Very Fast", "7: Super Fast", "8: Ultra Fast"}},
                         New NumParam With {.Switch = "-q", .Text = "QP", .Config = {0, 63, 1}, .Init = 50})
 
                     For Each item In ItemsValue